% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantETL.R
\name{getSNVobject}
\alias{getSNVobject}
\title{Make a SNV variant object from the returned variant data frame}
\usage{
getSNVobject(variantdf)
}
\arguments{
\item{variantdf}{}
}
\value{

}
\description{
Make a SNV variant object from the returned variant data frame
}
