% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantETL.R
\name{getVariants}
\alias{getVariants}
\title{Retrieve variants from the database}
\usage{
getVariants(
  tblname = "snvs",
  tblclass = "vdata",
  everything = F,
  qpjid = "0C734366DE0C36BBE56C1CF2FB74C24A",
  qgene = "TP53",
  qsample = "048_EK_8-308_S42",
  qtype = "sample",
  allcols = TRUE
)
}
\arguments{
\item{tblname}{}

\item{tblclass}{}

\item{qpjid}{}

\item{qgene}{}

\item{qsample}{}

\item{qtype}{}

\item{all}{}
}
\value{

}
\description{
Retrieve variants from the database
}
