% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantETL.R
\name{loadVariantBulk}
\alias{loadVariantBulk}
\title{Insert data from a file/cloud path into ES store}
\usage{
loadVariantBulk(
  pid = "pjx",
  tblclass = "vdata",
  skipLoad = FALSE,
  istart = 1,
  tblname = "snvs",
  infile = NULL,
  source = "cloud",
  s3path = mydf$prefixpath[1],
  s3bucket = NULL,
  doFilter = TRUE
)
}
\arguments{
\item{pid}{}

\item{tblclass}{}

\item{istart}{}

\item{tblname}{}

\item{infile}{}

\item{source}{}

\item{s3path}{}

\item{s3bucket}{}

\item{doFilter}{}

\item{mydf}{}
}
\value{

}
\description{
Insert data from a file/cloud path into ES store
}
